feat: moved tests to __test__ dir
